Fix message not accessing config provider
